# apps/admin-api/src/main/java/com/easy1auth/admin/web/AuthController.java
package com.easy1auth.admin.web;

import com.easy1auth.admin.config.AdminJwtProperties;
import com.easy1auth.admin.security.AdminTokenService;
import com.easy1auth.adminidentity.*;
import com.easy1auth.foundation.error.DomainException;
import com.easy1auth.tenant.TenantService;
import jakarta.servlet.http.HttpServletRequest;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.web.bind.annotation.*;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.mail.SimpleMailMessage;
import org.springframework.mail.javamail.JavaMailSender;
import com.easy1auth.adminaccess.AdminAccessService;
import com.easy1auth.admin.config.RegistrationProperties;
import com.easy1auth.security.SecurityPolicyService;
import com.easy1auth.audit.DeliveryService;
import org.springframework.security.oauth2.jwt.Jwt;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import java.util.*;

@RestController
@RequestMapping("/api/auth")
public class AuthController {
    private final AdminIdentityService identities; private final AdminTokenService tokens;
    private final TenantService tenants; private final AdminJwtProperties jwt;
    private final AdminAccessService access; private final RegistrationCodeService registrationCodes; private final JavaMailSender mail; private final RegistrationProperties registration; private final SecurityPolicyService security; private final DeliveryService delivery;
    AuthController(AdminIdentityService identities, AdminTokenService tokens, TenantService tenants, AdminJwtProperties jwt,AdminAccessService access,RegistrationCodeService registrationCodes,JavaMailSender mail,RegistrationProperties registration,SecurityPolicyService security,DeliveryService delivery) {
        this.identities=identities; this.tokens=tokens; this.tenants=tenants; this.jwt=jwt;this.access=access;this.registrationCodes=registrationCodes;this.mail=mail;this.registration=registration;this.security=security;this.delivery=delivery;
    }

    @PostMapping("/login")
    public LoginResponse login(@RequestBody LoginRequest request, HttpServletRequest http) {
        if (!"password".equals(request.loginType())) throw new DomainException("LOGIN_TYPE_UNSUPPORTED", "当前仅支持密码登录", 400);
        var result = identities.authenticate(request.username(), request.password(), jwt.refreshTtl(), http.getHeader("User-Agent"), http.getRemoteAddr());
        if(result.account().mfaEnabled()){identities.logout(result.refreshToken());var challenge=security.issueTotpChallenge("admin",result.account().id(),null,"login");return LoginResponse.mfa(challenge.token(),challenge.expiresIn());}
        return response(result.account(), result.refreshToken());
    }

    @PostMapping("/mfa/verify")
    public LoginResponse verifyMfa(@RequestBody MfaLoginRequest request,HttpServletRequest http){UUID account=security.consumeTotpChallenge(request.challengeToken(),request.code());var result=identities.completeMfa(account,jwt.refreshTtl(),http.getHeader("User-Agent"),http.getRemoteAddr());return response(result.account(),result.refreshToken());}

    @PostMapping("/register")
    @Transactional
    public LoginResponse register(@RequestBody RegisterRequest request, HttpServletRequest http) {
        String username = request.username() == null || request.username().isBlank() ? request.email().split("@",2)[0] : request.username();
        var result = identities.register(username, request.email(), request.password(), request.code(), jwt.refreshTtl(), http.getHeader("User-Agent"), http.getRemoteAddr());
        var tenant=tenants.create(result.account().id(),username+"的租户","basic");
        access.ensureDefaultRoles(tenant.id());
        return response(result.account(), result.refreshToken());
    }

    @PostMapping("/send-code") @Transactional public Map<String,Object> sendCode(@RequestBody SendCodeRequest request){if(!"register".equals(request.type()))throw new DomainException("CODE_TYPE_UNSUPPORTED","当前仅支持注册验证码",400);var issued=registrationCodes.issue(request.email());delivery.enqueueEmail(null,issued.email(),"Easy1Auth 注册验证码","您的验证码是 "+issued.code()+"，10分钟内有效。","registration:"+issued.email()+":"+java.time.Instant.now().getEpochSecond()/60);var body=new HashMap<String,Object>();body.put("status","success");body.put("message","验证码已进入发送队列");if(registration.exposeCode())body.put("code",issued.code());return body;}

    @PostMapping("/refresh")
    public Map<String,String> refresh(@RequestBody RefreshRequest request, HttpServletRequest http) {
        var result = identities.rotate(request.refreshToken(), jwt.refreshTtl(), http.getHeader("User-Agent"), http.getRemoteAddr());
        return Map.of("token", tokens.issue(result.account()), "refreshToken", result.replacementToken());
    }

    @PostMapping("/logout")
    public Map<String,String> logout(@RequestBody(required=false) RefreshRequest request,@AuthenticationPrincipal Jwt principal) {
        if(request!=null&&request.refreshToken()!=null)identities.logout(request.refreshToken());else identities.logoutAllAndInvalidate(UUID.fromString(principal.getSubject()));
        return Map.of("status", "success", "message", "退出成功");
    }

    private LoginResponse response(AdminAccount account, String refresh) {
        var memberships = tenants.list(account.id());
        UUID current = account.lastTenantId();
        boolean currentAccessible = current != null && memberships.stream().map(t -> t.id()).anyMatch(current::equals);
        if (!currentAccessible) current = memberships.isEmpty() ? null : memberships.getFirst().id();
        return LoginResponse.success(tokens.issue(account), refresh,new LoginUser(account.id(), account.username(), account.email(), null, current), memberships);
    }
    public record LoginRequest(String username, String password, String email, String code, String loginType) {}
    public record RegisterRequest(String email, String password, String code, String username) {}
    public record RefreshRequest(String refreshToken) {}
    public record SendCodeRequest(String email,String type){}
    public record MfaLoginRequest(String challengeToken,String code){}
    public record LoginUser(UUID id, String username, String email, String avatar, UUID currentTenantId) {}
    public record LoginResponse(String status,String token,String refreshToken,LoginUser user,List<?> tenants,String challengeToken,List<String> methods,Integer expiresIn) {
        static LoginResponse success(String token,String refresh,LoginUser user,List<?> tenants){return new LoginResponse("success",token,refresh,user,tenants,null,null,null);}
        static LoginResponse mfa(String challenge,int expires){return new LoginResponse("mfa_required",null,null,null,List.of(),challenge,List.of("totp"),expires);}
    }
}