# docs/java-migration/admin-api-openapi.yaml
openapi: 3.1.0
info:
  title: Easy1Auth Admin API compatibility contract
  version: 0.1.0
servers:
  - url: /api
components:
  securitySchemes:
    adminBearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    tenantId:
      name: tenant-id
      in: header
      required: true
      description: Selected tenant; server must independently validate active membership.
      schema: { type: string, format: uuid }
  schemas:
    Error:
      type: object
      required: [status, message]
      properties:
        status: { type: string, enum: [fail, error] }
        code: { type: string }
        message: { type: string }
    Tenant:
      type: object
      required: [id, name, status, plan, role]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string }
        plan: { type: string }
        role: { type: string }
security: [{ adminBearer: [] }]
paths:
  /auth/login:
    post:
      security: []
      summary: Administrator login (phase 3)
      responses:
        '200': { description: Legacy-compatible token and tenant response }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /auth/refresh:
    post:
      security: []
      summary: Rotate an opaque administrator refresh token
      responses:
        '200': { description: New access token and replacement refresh token }
        '401': { description: Expired, revoked, replayed, or stale-security-version session }
  /admin-users:
    get:
      summary: List memberships in the selected tenant
      parameters: [{ $ref: '#/components/parameters/tenantId' }]
      responses:
        '200': { description: Legacy-compatible admins pagination object }
        '403': { description: Membership or permission denied }
  /admin-roles:
    get:
      summary: List membership-scoped administrator roles
      parameters: [{ $ref: '#/components/parameters/tenantId' }]
      responses:
        '200': { description: Legacy-compatible roles pagination object }
        '403': { description: Membership or permission denied }
  /tenants/list:
    get:
      summary: List only tenants joined by the authenticated account
      responses:
        '200': { description: Legacy top-level tenants/page/pageSize/total response }
  /tenants/create:
    post:
      summary: Create tenant and owner membership atomically
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1 }
                plan: { type: string, default: basic }
      responses:
        '200': { description: Legacy top-level tenant response }
        '400': { description: Validation failure, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /tenants/current:
    get:
      summary: Resolve selected tenant through membership
      parameters: [{ $ref: '#/components/parameters/tenantId' }]
      responses:
        '200': { description: Current tenant }
        '403': { description: Missing/inactive membership, account, or tenant }
  /tenants/{tenantId}/owner-transfer:
    post:
      summary: Transfer the single owner under a tenant row lock
      parameters:
        - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Transferred }
        '409': { description: Ownership invariant conflict }